pollsapi/requirement.txt

Django==2.2.6
django-admin-generator==2.2.0
django-seed==0.1.9
djangorestframework==3.10.3
drf-dynamic-fields==0.3.1


